requirements.txt
streamlit
openai
stripe
python-dotenv
pyinstaller
